+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - Unreleased
+
+### Added
+
+- Added overflow methods
+- Added no_wrap option to print()
+
+### Fixed
+
+- Improved handling of compressed tables
+
+### Changed
+
+- Renamed \_ratio.ratio_divide to \_ratio.ratio_distribute
+
 ## [1.3.1] - 2020-06-01
 
 ### Changed